==> Transfers_BR/GK/01-GK.ipynb <==
import pandas as pd
from bs4 import BeautifulSoup
from datetime import datetime


#TIMESTAMP DO COMEÇO
hora_atual = datetime.now()
hora_formatada = hora_atual.strftime("%H:%M:%S")
print("Hora Inicial:", hora_formatada)

arquivos = ['GK_2014_1.html', 'GK_2014_2.html', 'GK_2015_1.html', 'GK_2015_2.html', 'GK_2016_1.html', 'GK_2016_2.html', 'GK_2017_1.html', 'GK_2017_2.html', 'GK_2018_1.html', 'GK_2018_2.html', 'GK_2019_1.html', 'GK_2019_2.html', 'GK_2020_1.html', 'GK_2020_2.html', 'GK_2021_1.html', 'GK_2021_2.html', 'GK_2022_1.html', 'GK_2022_2.html', 'GK_2023_1.html', 'GK_2023_2.html', 'GK_2024_1.html', 'GK_2024_2.html', 'GK_2025_1.html', 'GK_2025_2.html']

# Inicialize um DataFrame principal
df_principal = pd.DataFrame()

for arquivo in arquivos:
    try:
        with open(arquivo, 'r', encoding='utf-8') as file:
            html_content = file.read()
    except FileNotFoundError:
        print(f"Arquivo não encontrado: {arquivo}")
        continue

    soup = BeautifulSoup(html_content, 'html.parser')
    table = soup.find('table', {'class': 'items'})
    
    if not table:
        print(f"Tabela não encontrada no arquivo: {arquivo}")
        continue

    rows = table.find_all('tr')
    data = []
    
    for row in rows[:]:
        cols = row.find_all('td')
        row_data = []
        for col in cols:
            link = col.find('a')
            if link:
                row_data.append(link['href'])
                row_data.append(link.text.strip())
            elif col.find('img'):
                img_alt = col.find('img')['alt']
                row_data.append(img_alt.strip())
            else:
                row_data.append(col.text.strip())

        #if len(row_data) == len(columns):
        # Adicionar a linha, mesmo que o número de colunas seja diferente do esperado
        if len(row_data) > 5:  # Apenas adicionar se houver dados
            data.append(row_data)
            

    columns = [
        'ID', 'Perfil', 'Nome', 'Nome Completo', 'URL Nome Completo', 'Repetido Nome', 'Posição',
        'Idade', 'Valor de Mercado', 'Nacionalidade', 'URL Clube Antigo', 'Vazio 1', 'URL Clube Antigo Repetido',
        'Vazio 2', 'URL Clube Antigo Repetido 2', 'Clube Antigo', 'URL Liga Antiga', 'Liga Antiga',
        'URL Clube Novo', 'Vazio 3', 'URL Clube Novo Repetido', 'Vazio 4', 'URL Clube Novo Repetido 2',
        'Clube Novo', 'URL Liga Nova', 'Liga Nova', 'Valor Transferência'
    ]
    
    df = pd.DataFrame(data)
    
    # Adicionar ao DataFrame principal
    df_principal = pd.concat([df_principal, df], ignore_index=True)
df_principal

# Salvar em CSV
df_principal.to_csv('01-GK.csv.csv', index=False)

#TIMESTAMP DO FINAL
hora_atual = datetime.now()
hora_formatada = hora_atual.strftime("%H:%M:%S")
print("Hora Final:", hora_formatada)
