==> Transfers_BR/GK/04- GK.ipynb <==
# Import/load

import pandas as pd
import numpy as np

df_base = pd.read_csv('02-GK.csv')
df_home = pd.read_csv('03-GK-home.csv')
df_stats = pd.read_csv('03-GK-stats.csv')

# head

df_base.head(3).T

df_home.head(3).T

df_stats.head(3).T

# Checagem de tipo

print("df_base:",df_base.shape)
print("df_home:",df_home.shape)
print("df_stats:",df_stats.shape)

df_base.dtypes

df_home.dtypes

df_stats.dtypes

# Descarte de colunas

df_home = df_home.drop(columns=["Unnamed: 0"])
df_home = df_home.drop(columns=["nome_hifen"])


df_stats = df_stats.drop(columns=["Unnamed: 0"])

# Merges

df_merged = df_base.merge(df_home, on="ID", how="inner")
print("df_merged:",df_merged.shape)

df_merged.sample(3).T

# Cria uma nova coluna em df_stats com ano_stat + 1
df_stats["ano_transferencia"] = df_stats["ano_stat"] + 1

# Faz o merge com base em ID e no ano ajustado
df_final = df_merged.merge(
    df_stats,
    left_on=["ID", "Ano da Tranferência"],
    right_on=["ID", "ano_transferencia"],
    how="left"
)

# Remove a coluna temporária "ano_transferencia" se não for mais necessária
df_final = df_final.drop(columns=["ano_transferencia"])
print("df_final:",df_final.shape)

# Limpeza e tipagem

print(df_final.columns)  # Listar todas as colunas para conferir se tudo está certo

print(df_final.sample(1).T)  # Ver uma linha aleatória

print(df_final.isnull().sum())  # Verificar se há valores NaN após o merg

print(df_final.dtypes)

df_final.head(3).T

#df_final.replace("-", pd.NA, inplace=True)

df_final["altura"] = df_final["altura"].str.replace(" m", "").str.replace(",", ".").astype(float)

# Função para converter valores no formato '1.025'
def convert_minutes(val):
    if isinstance(val, str) and val.endswith("'"):
        # Remover o apóstrofo e multiplicar por 1000
        return int(float(val.replace("'", "")) * 1000)
    return val  # Se não for uma string no formato, retorna o valor original

# Aplicando a função nas colunas 'minutos' e 'minutos_por_gol'
df_final['minutos_por_gol'] = df_final['minutos_por_gol'].apply(convert_minutes)
df_final['minutos'] = df_final['minutos'].apply(convert_minutes)

df_final["Valor de Mercado"] = df_final["Valor de Mercado"].astype("int64")
df_final["Valor Transferência"] = df_final["Valor Transferência"].astype("int64")

df_final.dtypes

df_final.head(6).T

# Substituir '-' por NaN
cols_to_clean = [
    "jogou", "gols_para", "assistencias", "gols_contra", 
    "entrou", "saiu", "amarelo", "amarelo_duplo", 
    "vermelho", "Gols_de_penalti", "minutos_por_gol", "minutos"
]

df_final[cols_to_clean] = df_final[cols_to_clean].replace('-', pd.NA)
# Convertemos para inteiros, substituindo NaN por 0 (se necessário)
#df_final[cols_to_clean] = df_final[cols_to_clean].fillna(0).astype("int64")


df_final.shape

num_duplicatas = df_final.duplicated().sum()
print(f"Número de linhas duplicadas: {num_duplicatas}")

df_final.head(6).T

df_final['minutos'] = df_final['minutos'].apply(lambda x: str(x)[:-3] if str(x).endswith('000') else x)

# Renomear Colunas

# Renomeando as colunas
df_final.rename(columns={
    'Gols_de_penalti': 'gols_sofridos',
    'minutos_por_gol': 'clean_sheets'
}, inplace=True)

# Verificando se a renomeação foi realizada com sucesso
print(df_final.columns)

# Salvo

df_final=df_final.drop_duplicates()
df_final.to_csv('04-GK.csv', index=True)

df_final.shape

df_final.dtypes
